♻️refactor: Status enum을 global 패키지로 이동
