+++ src/main/java/com/project/likelion13th_team1/domain/event/entity/Event.java
@@ -1,7 +1,7 @@
 package com.project.likelion13th_team1.domain.event.entity;
 
 import com.project.likelion13th_team1.domain.routine.entity.Routine;
-import com.project.likelion13th_team1.domain.routine.entity.Status;
+import com.project.likelion13th_team1.global.entity.Status;
 import jakarta.persistence.*;
 import lombok.*;
