+++ src/main/java/com/project/likelion13th_team1/global/entity/Status.java
@@ -1,4 +1,4 @@
-package com.project.likelion13th_team1.domain.routine.entity;
+package com.project.likelion13th_team1.global.entity;
 
 public enum Status {
     // 진행 중, 성공, 취소
